chore: Update index.md and _config.yml content
Updated the 'About Me' section content and format in index.md to provide a more concise and appealing description. The method for contact has been restructured and an iframe for bookings inserted. Changes are matched in _config.yml with an addition of 'Contact' in the main navigation. A 'footer-navigation.html' attribute was removed from _config.yml.